backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
motor==3.3.1
pymongo==4.5.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.0
pytest==7.4.3
pytest-asyncio==0.21.1
